refactor: 상품 모달 코드 구조 개선
- styled 모듈화
