+++ src/components/ProductModalContent/index.tsx
@@ -1,0 +1,3 @@
+import ProductModalContent from "./ProductModalContent";
+
+export default ProductModalContent;
